requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0  # For Excel file reading

# Text processing
nltk>=3.8
scikit-learn>=1.3.0
unicodedata2>=15.0.0  # Better Unicode normalization

# Visualization
matplotlib>=3.7.0
wordcloud>=1.9.0
networkx>=3.1

# Statistical analysis
scipy>=1.10.0
statsmodels>=0.14.0

# YAML configuration
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.7.0
flake8>=6.1.0
